# cpython: enable conservative performance improvement that doesn't harm reproducibility