@raboof 2021-05-25
> Another person (preferably on an Intel CPU) should reproduce the Python build (twice)

I built f0430aab956 with `nix-build -A python3 -I nixpkgs=/home/aengelen/nixpkgs-r13y --option sandbox true --check --keep-failed` and that successfully produced /nix/store/5hdg7s35lgvndaii8jigi1b2mjhrwl11-python3-3.8.9 with an Intel i9-9980HK. `nix-hash /nix/store/5hdg7s35lgvndaii8jigi1b2mjhrwl11-python3-3.8.9` yields `879718710e16b22c4f2f0974e5f17cf1`