@mweinelt 2021-05-25
```
❯ nom-build -A python3 --check
/nix/store/5hdg7s35lgvndaii8jigi1b2mjhrwl11-python3-3.8.9
❯ nix-hash /nix/store/5hdg7s35lgvndaii8jigi1b2mjhrwl11-python3-3.8.9
879718710e16b22c4f2f0974e5f17cf1
```