@sternenseemann 2021-05-26
LTO usually involves compiler-specific bintools to work, so we should make sure all required are available with all our stdenvs. The linker is usually not a problem because we load a plugin for that (i. e. clang LTO works with the GNU linker), but instead of the respective default `ar` we need to make sure we can use `gcc-ar` and `llvm-ar` respectively.

* For gcc this is not complicated at all: `gcc.out` or rather `stdenv.cc.cc.out` contains `gcc-ar` etc. in its `bin` directory and should be in `PATH` by default. cpython doesn't seem to care about `gcc-ar` though, it either just works or we'll have to override that.
* For clang it is a bit more complicated because we have different kinds of stdenvs using it and `llvm-ar` belongs to the llvm package (or our `bintools` pseudo package which symlinks to llvm).
    * If `stdenv.hostPlatform.linker == "lld"` `stdenv.cc.bintools.bintools` will contain `llvm-ar` etc. and should be in `PATH` automatically. This should is the case for e. g. `pkgsLLVM`
    * In all other cases we need to get `llvm-ar` etc. from llvm somehow. `clang-unwrapped` exposes the llvm version it uses via a passthru attribute, so we'd need to add `stdenv.cc.cc.libllvm.out` to `nativeBuildInputs` or equivalent. This case should be matched by `stdenv.hostPlatform.linker != "lld" && stdenv.cc.isClang`.

I can (sanity) check compilation in `pkgsLLVM` for the `useLLVM` stdenv, but would expect it to work. Darwin may be a bit of trouble and using `llvmPackages.stdenv`.