@mweinelt 2021-06-17
RaspberryPi 4

```
[root@dione:/etc/nixos/nixpkgs]# nix-build -A python38
/nix/store/5z6lghw3lv7h2svm0bmrrvkayvp2ibmc-python3-3.8.9
[root@dione:/etc/nixos/nixpkgs]# nix-hash /nix/store/5z6lghw3lv7h2svm0bmrrvkayvp2ibmc-python3-3.8.9
a4090f468533d4f8f6bae43fa15fd072
```

Oracle Cloud Ampere A1 VM

```
[root@pallas:/etc/nixos/nixpkgs]# nix-build -A python38
/nix/store/5z6lghw3lv7h2svm0bmrrvkayvp2ibmc-python3-3.8.9
[root@pallas:/etc/nixpkgs]# nix-hash /nix/store/5z6lghw3lv7h2svm0bmrrvkayvp2ibmc-python3-3.8.9
a4090f468533d4f8f6bae43fa15fd072
```