@SuperSandro2000 2021-06-18
```ShellSession
$ nix-build -A python37 -A python38 -A python39 -A python310
/nix/store/8hgvwjr17pn8m4965d9r130ilxdgb485-python3-3.7.10
/nix/store/hd3pnn3ykgnrkm2dc3wfg4m498xcv7lc-python3-3.8.9
/nix/store/1462shpz72qkv57xxxvii4svq0fdzh1x-python3-3.9.4
/nix/store/frz5vf2pvryp9v33hbn02ni2lagc5vsv-python3-3.10.0a5

$ nix-hash /nix/store/8hgvwjr17pn8m4965d9r130ilxdgb485-python3-3.7.10 /nix/store/hd3pnn3ykgnrkm2dc3wfg4m498xcv7lc-python3-3.8.9 /nix/store/1462shpz72qkv57xxxvii4svq0fdzh1x-p
ython3-3.9.4 /nix/store/frz5vf2pvryp9v33hbn02ni2lagc5vsv-python3-3.10.0a5334499c2033b665f165aa210f5115e76
c0b091771c6dc058b92e5004e594cbe9
2ccf8ab7f9521293de1e664ac0a77454
c4334e81a2e1719374d7e1c1363b828c
```