@mweinelt 2021-06-18
Picked on top of 43362bd872ee56f7c3c391e84d7bfaf900de7f93 on both machines. The idea was to check whether the output is reproducible between two aarch64 hosts.