@SuperSandro2000 2021-06-18
That produces /nix/store/pd1r11g83h8axp7fkv6yhbq1m9ibs1dh-python3-3.8.9 everytime for me