@mweinelt 2021-06-29
Fixed up the typo and rebased onto python-unstable.